requirements.txt
underthesea>=1.3.6
stanza
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
transformers>=4.40.0
datasets>=2.18.0
torch>=2.0.0
fastembed
faiss-cpu>=1.7.4
openai
streamlit
python-dotenv
rank_bm25
requests
sentence-transformers>=3.0.0,<4.0.0